win: build fixes
With https://codereview.chromium.org/709793002/ for base/, build/, gpu/, 'mojo' builds in mojo again.

R=jam@chromium.org

Review URL: https://codereview.chromium.org/708133002

+++ mojo/shell/dynamic_application_loader.cc
@@ -85,7 +85,7 @@ class DynamicApplicationLoader::Loader {
     if (PeekContentHandler(&shebang, &url)) {
       load_callbacks_->LoadWithContentHandler(
           url, AsURLResponse(context_->task_runners()->blocking_pool(),
-                             shebang.size()));
+                             static_cast<int>(shebang.size())));
       return;
     }
